22 backend integration

+++ back/app/Http/Controllers/TicketController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\OwesPayment;
 use App\Models\Ticket;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TicketController extends Controller
 {
@@ -39,16 +41,28 @@ class TicketController extends Controller
     {
         $request->validate([
             'name' => 'required|string',
-            'amount' => 'required|numeric'
+            'amount' => 'required|numeric',
+            'expiration_date' => 'nullable|date'
         ]);
 
         $user = Auth::user();
+
         $ticket = Ticket::create([
             'name' => $request->name,
             'amount' => $request->amount,
+            'expiration_date' => $request->expiration_date ? date("Y-m-d H:i", strtotime(date($request->expiration_date))) : null,
             'creator_id' => $user->id
         ]);
         $ticket['creator'] = $ticket->creator;
+
+        $residents = $user->appartment->residents;
+        foreach ($residents as $resident) {
+            OwesPayment::create([
+                'requested_amount' => $request->amount / count($residents),
+                'for_ticket_id' => $ticket->id,
+                'payer_id' => $resident->id
+            ]);
+        }
 
         return response()->json([
             'ticket' => $ticket
@@ -58,9 +72,11 @@ class TicketController extends Controller
     public function deleteTicket(Request $request, $ticket_id)
     {
         $ticket = Ticket::find($ticket_id);
+
+        Gate::authorize('admin-or-owner', [$ticket, $ticket->creator->appartment, 'creator_id']);
+
         $ticket->delete();
 
         return response()->json([], 204);
     }
 }
-
